+++ config/config.go
@@ -18,7 +18,7 @@ type Configurations struct {
 
 // reference: https://stackoverflow.com/questions/63889004/how-to-access-specific-items-in-an-array-from-viper
 type ClusterListConfiguration struct {
-	Name                    string    `mapstructure:"Name"`
+	ClusterName             string    `mapstructure:"ClusterName"`
 	AwsRegion               string    `mapstructure:"AwsRegion"`
 	AwsAccount              string    `mapstructure:"AwsAccount"`
 	AwsNodeObject           K8sObject `mapstructure:"AwsNodeObject"`
@@ -28,8 +28,10 @@ type ClusterListConfiguration struct {
 }
 
 type K8sObject struct {
-	Name string `mapstructure:"name"`
-	Type string `mapstructure:"type"`
+	DeploymentName string `mapstructure:"DeploymentName"`
+	ObjectType     string `mapstructure:"ObjectType"`
+	ContainerName  string `mapstructure:"ContainerName"`
+	Namespace      string `mapstructure:"Namespace"`
 }
 
 type ComponentVersionConfigurations struct {
@@ -39,11 +41,24 @@ type ComponentVersionConfigurations struct {
 	KubeProxy         string `mapstructure:"kube-proxy"`
 }
 
-// TODO Add spec to check for unique cluster name list in config
 func (c Configurations) IsClusterListConfigurationValid() bool {
 	valid := true
+	clusterNameMap := map[string]string{}
 	for _, cluster := range c.ClusterList {
-		if cluster.Name == "" || cluster.AwsRegion == "" || cluster.AwsAccount == "" || cluster.AwsNodeObject.Name == "" || cluster.AwsNodeObject.Type == "" || cluster.ClusterAutoscalerObject.Name == "" || cluster.ClusterAutoscalerObject.Type == "" || cluster.CoreDnsObject.Name == "" || cluster.CoreDnsObject.Type == "" || cluster.KubeProxyObject.Name == "" || cluster.KubeProxyObject.Type == "" {
+		if _, present := clusterNameMap[cluster.ClusterName]; present {
+			valid = false
+		}
+		clusterNameMap[cluster.ClusterName] = "present"
+
+		clusterName := cluster.ClusterName == ""
+		awsRegion := cluster.AwsRegion == ""
+		awsAccount := cluster.AwsAccount == ""
+		awsNodeObject := cluster.AwsNodeObject.DeploymentName == "" || cluster.AwsNodeObject.ObjectType == "" || cluster.AwsNodeObject.ContainerName == "" || cluster.AwsNodeObject.Namespace == ""
+		clusterAutoscaler := cluster.ClusterAutoscalerObject.DeploymentName == "" || cluster.ClusterAutoscalerObject.ObjectType == "" || cluster.ClusterAutoscalerObject.ContainerName == "" || cluster.ClusterAutoscalerObject.Namespace == ""
+		coreDns := cluster.CoreDnsObject.DeploymentName == "" || cluster.CoreDnsObject.ObjectType == "" || cluster.CoreDnsObject.ContainerName == "" || cluster.CoreDnsObject.Namespace == ""
+		kubeProxy := cluster.KubeProxyObject.DeploymentName == "" || cluster.KubeProxyObject.ObjectType == "" || cluster.KubeProxyObject.ContainerName == "" || cluster.KubeProxyObject.Namespace == ""
+
+		if clusterName || awsRegion || awsAccount || awsNodeObject || clusterAutoscaler || coreDns || kubeProxy {
 			valid = false
 		}
 	}
@@ -61,36 +76,66 @@ func (c Configurations) IsComponentVersionConfigurationsValid() bool {
 func (c Configurations) IsClusterNameValid(clusterName string) bool {
 	contains := false
 	for _, cluster := range c.ClusterList {
-		if cluster.Name == clusterName {
+		if cluster.ClusterName == clusterName {
 			contains = true
 		}
 	}
 	return contains
 }
 
-func (c Configurations) GetK8sObjectNameAndObjectTypeForCluster(clusterName, k8sObject string) (objectName, objectType string, err error) {
+func (c Configurations) GetK8sObjectForCluster(clusterName, k8sObjectType string) (k8sObject K8sObject, err error) {
 	for _, cluster := range c.ClusterList {
-		if cluster.Name == clusterName {
-			switch k8sObject {
+		if cluster.ClusterName == clusterName {
+			switch k8sObjectType {
 			case "aws-node":
-				return cluster.AwsNodeObject.Name, cluster.AwsNodeObject.Type, nil
+				return K8sObject{
+					DeploymentName: cluster.AwsNodeObject.DeploymentName,
+					ObjectType:     cluster.AwsNodeObject.ObjectType,
+					ContainerName:  cluster.AwsNodeObject.ContainerName,
+					Namespace:      cluster.AwsNodeObject.Namespace,
+				}, nil
 			case "cluster-autoscaler":
-				return cluster.ClusterAutoscalerObject.Name, cluster.ClusterAutoscalerObject.Type, nil
+				return K8sObject{
+					DeploymentName: cluster.ClusterAutoscalerObject.DeploymentName,
+					ObjectType:     cluster.ClusterAutoscalerObject.ObjectType,
+					ContainerName:  cluster.ClusterAutoscalerObject.ContainerName,
+					Namespace:      cluster.ClusterAutoscalerObject.Namespace,
+				}, nil
 			case "kube-proxy":
-				return cluster.KubeProxyObject.Name, cluster.KubeProxyObject.Type, nil
+				return K8sObject{
+					DeploymentName: cluster.KubeProxyObject.DeploymentName,
+					ObjectType:     cluster.KubeProxyObject.ObjectType,
+					ContainerName:  cluster.KubeProxyObject.ContainerName,
+					Namespace:      cluster.KubeProxyObject.Namespace,
+				}, nil
 			case "coredns":
-				return cluster.CoreDnsObject.Name, cluster.CoreDnsObject.Type, nil
+				return K8sObject{
+					DeploymentName: cluster.CoreDnsObject.DeploymentName,
+					ObjectType:     cluster.CoreDnsObject.ObjectType,
+					ContainerName:  cluster.CoreDnsObject.ContainerName,
+					Namespace:      cluster.CoreDnsObject.Namespace,
+				}, nil
 			default:
-				return "", "", errors.New("please pass any of the components between aws-node, coredns, cluster-autoscaler, kube-proxy")
+				return K8sObject{
+					DeploymentName: "",
+					ObjectType:     "",
+					ContainerName:  "",
+					Namespace:      "",
+				}, errors.New("please pass any of the components between aws-node, coredns, cluster-autoscaler, kube-proxy")
 			}
 		}
 	}
-	return "", "", errors.New("please check if you passed a valid cluster name")
+	return K8sObject{
+		DeploymentName: "",
+		ObjectType:     "",
+		ContainerName:  "",
+		Namespace:      "",
+	}, errors.New("please check if you passed a valid cluster name")
 }
 
 func (c Configurations) GetAwsAccountAndRegionForCluster(clusterName string) (awsAccount, awsRegion string, err error) {
 	for _, cluster := range c.ClusterList {
-		if cluster.Name == clusterName {
+		if cluster.ClusterName == clusterName {
 			return cluster.AwsAccount, cluster.AwsRegion, nil
 		}
 	}
@@ -146,7 +191,7 @@ func Read(fileName, fileType, filePath string) (config Configurations, err error
 	}
 
 	if !config.IsClusterListConfigurationValid() {
-		return Configurations{}, errors.New("one of the clusterlist elements has either Name, AwsRegion, AwsAccount, AwsNodeObject, ClusterAutoscalerObject, KubeProxyObject, CoreDnsObject is missing")
+		return Configurations{}, errors.New("one of the clusterlist elements has either ClusterName, AwsRegion, AwsAccount, AwsNodeObject, ClusterAutoscalerObject, KubeProxyObject, CoreDnsObject is missing")
 	}
 
 	return config, nil